% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwiser.R
\docType{package}
\name{pwiser-package}
\alias{pwiser}
\alias{pwiser-package}
\title{pwiser: Tidy P(air)wise Operations}
\description{
To learn more about pwiser see https://github.com/brshallo/pwiser
}
\author{
\strong{Maintainer}: Bryan Shalloway \email{brshallodev@gmail.com}

Authors:
\itemize{
  \item Carlo Medina
}

}
\keyword{internal}
